Add unit tests and do some cleanup
