+++ packages/react-router-dom/__tests__/data-browser-router-test.tsx
@@ -6,7 +6,6 @@ import {
   fireEvent,
   waitFor,
   screen,
-  prettyDOM,
 } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import type { ErrorResponse, Fetcher } from "@remix-run/router";
@@ -36,6 +35,9 @@ import {
   useSearchParams,
   createRoutesFromElements,
 } from "react-router-dom";
+
+import createDeferred from "../../router/__tests__/utils/createDeferred";
+import getHtml from "../../react-router/__tests__/utils/getHtml";
 
 testDomRouter("<DataBrowserRouter>", createBrowserRouter, (url) =>
   getWindowImpl(url, false)
@@ -5925,48 +5927,9 @@ function testDomRouter(
   });
 }
 
-function createDeferred() {
-  let resolve: (val?: any) => Promise<void>;
-  let reject: (error?: Error) => Promise<void>;
-  let promise = new Promise((res, rej) => {
-    resolve = async (val: any) => {
-      res(val);
-      try {
-        await promise;
-      } catch (e) {}
-    };
-    reject = async (error?: Error) => {
-      rej(error);
-      try {
-        await promise;
-      } catch (e) {}
-    };
-  });
-  return {
-    promise,
-    //@ts-ignore
-    resolve,
-    //@ts-ignore
-    reject,
-  };
-}
-
 function getWindowImpl(initialUrl: string, isHash = false): Window {
   // Need to use our own custom DOM in order to get a working history
   const dom = new JSDOM(`<!DOCTYPE html>`, { url: "http://localhost/" });
   dom.window.history.replaceState(null, "", (isHash ? "#" : "") + initialUrl);
   return dom.window as unknown as Window;
-}
-
-function getHtml(container: HTMLElement) {
-  return prettyDOM(container, undefined, {
-    highlight: false,
-    theme: {
-      comment: null,
-      content: null,
-      prop: null,
-      tag: null,
-      value: null,
-    },
-  });
 }
