+++ packages/react-router/__tests__/utils/getHtml.ts
@@ -1,0 +1,7 @@
+import { prettyDOM } from "@testing-library/react";
+
+export default function getHtml(container: HTMLElement) {
+  return prettyDOM(container, undefined, {
+    highlight: false,
+  });
+}
