+++ packages/react-router/__tests__/utils/tick.ts
@@ -1,0 +1,3 @@
+export default async function tick() {
+  await new Promise((r) => setTimeout(r, 0));
+}
